Update PR from advice in comments
