+++ samples/snippets/requirements-test.txt
@@ -1,0 +1,2 @@
+google-auth==1.34.0
+pytest==6.2.4
