+++ samples/snippets/requirements.txt
@@ -1,0 +1,2 @@
+google-api-core==1.31.1
+google-cloud-contact-center-insights==0.2.0
