+++ samples/snippets/test_get_operation.py
@@ -13,14 +13,21 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-# [START contactcenterinsights_get_operation]
-from google.cloud.contact_center_insights_v1.services.contact_center_insights import client
-from google.longrunning import operations_pb2
+import google.auth
+
+import get_operation
 
 
-def get_operation(project_id: str, location_id: str, operation_id: str) -> operations_pb2.Operation:
-    insights_client = client.ContactCenterInsightsClient()
-    operation_name = f"projects/{project_id}/locations/{location_id}/operations/{operation_id}"
-    return insights_client.transport.operations_client.get_operation(operation_name)
+def test_get_operation(capsys):
+    _, project_id = google.auth.default()
+    location_id = "us-central1"
+    operation_id = "12345"
 
-# [END contactcenterinsights_get_operation]
+    try:
+        operation = get_operation.get_operation(project_id, location_id, operation_id)
+        assert operation.name == f"projects/{project_id}/locations/{location_id}/operations/{operation_id}"
+    except Exception as e:
+        if "not found" in str(e):
+            pass
+        else:
+            raise
